requirements.txt
numpy>=1.24.0
scipy>=1.11.0
Pillow>=10.0.0
pytest>=7.4.0

fastapi
uvicorn
httpx
